tools/create-image.sh: add linux-firmware package by default
Multiple usb drivers (among others) are missing firmware thus
stopping short any probe() attempts.

A couple examples are:
1) comedi usb drivers: usbdux, usbduxfast etc.
2) drivers/media/usb/{s2255,go7007}
and a few others at the very least.

Add the required module 'firmware-linux' (carries both free and
nonfree batches of built firmware) to the list of default packages.

This is not an ideal solution considering the goal of a *minimal*
distribution. Still, this will increase code coverage.