[WIP] fuzzer: attach security contexts to programs
Added two new experimental flags that allow the fuzzer to attach
simple security contexts to each program that will be executed on the
target. The objective is to allow syzkaller to fuzz syscalls under
specific security contexts that match the SELinux policy that is loaded
on target.

The `enforce_policy` flag may be used to enforce the policy before the
executor forks and calls the `execute_one()` function. The default
Debian security policy does not allow the dynamic transition to
`user_u:user_r:user_t:s0` through `setcon()`.

Signed-off-by: Rares Constantin <rarescon@google.com>