[WIP] fuzzer: attach security contexts to programs
Added one experimental field that allow the user to add a list of
security contexts that can be used during fuzzing. The security context
is attached to the program as the first "syscall"  and it will be set
using `setcon`. The objective is to allow syzkaller to fuzz syscalls
under specific security contexts that match the SELinux policy that
is loaded on target.

Signed-off-by: Rares Constantin <rarescon@google.com>