[WIP] fuzzer: add security context generator stub
Added the security context generator in the fuzzer package which will
contain the implementation for loading a binary security policy, parse
it and then generate an appropiate security context based on the
syscalls that are fuzzed in a program.

For now, it will use the first security context provided in the
`seccontexts` list if any is available.

Signed-off-by: Rares Constantin <rarescon@google.com>