vm: add context to Pool.Create()
Enable external abortion of the instance creation process. This is
especially useful for the qemu case where we retry the creation/boot up
to 1000 times, which can take significant time (e.g. it timeouts
syz-cluster pods on unstable kernels).

The context can be further propagated to WaitForSSH, but that requires
another quite significant vm/ refactoring.