tools/syz-kcidb: add -input and -output flags for testing
Add -input and -output flags to syz-kcidb to allow for faster and
easier testing of KCIDB submissions.

With these flags, it's possible to use a local JSON file as input
for a bug report and to save the resulting KCIDB submission locally.
This removes the need for communication with the dashboard and KCIDB,
which simplifies testing and development.

Signed-off-by: Denys Fedoryshchenko <denys.f@collabora.com>