pkg/email/lore: demand patches in patch series emails
On LKML, users sometimes reply to a patch series while keeping the
original title. At the same time, sometimes new series are sent as a
reply to the previous versions. It all confuses the lore archive
parsing logic.

When reading the archive, remember if the patch were present. If not,
mark the series as corrupted and thus prevent is appearance in
syz-cluster.