executor: determine the next free loop device
When setting up loop devices, syz_mount_image() and reset_loop() use
"/dev/loop/<<procid>>" as the loop device name.  The variable <<procid>>
is always 0 (it is never initialized) and so the loop device is always
"/dev/loop0" (see [*] for example).

On many systems, "/dev/loop0" doesn't exist a priori and loop devices
must be explicitly created.  Fix executor/common_linux.h to do just
that.

I have not faced a C reproducer that included syz_read_part_table(), so
the changes to that function are purely speculative.  Please verify.

[*] https://syzkaller.appspot.com/bug?extid=fa7122891ab9e0bbc6a7