pkg/csource: run testPseudoSyscalls() for all targets
Previously in the -short configuration testPseudoSyscalls() was only
executed for the first supported target, which in most cases ended up
being linux/386 (the least popular configuration).
As a result, platform-specific pseudo-syscalls were never executed on
the CI.

Fix this by testing the pseudo-syscalls on all available platforms.
This increases the execution time of TestGenerate from 18s to 36s,
but also helps to discover bugs in pseudo-syscalls quicker.

As a result of this change, 3 distinct latent bugs were found on
amd64, arm64 and ppc64.