tools/create-image.sh: add option to specify output prefix
Add an option, -o / --output, to specify a prefix used for the name of the
directory where debootstrap generates the system, and the final disk image
and SSH key filenames.

The default remains using the distro release codename.

For now, ban the use of slashes, spaces, . and .. as output names.

Signed-off-by: Andrew Donnellan <ajd@linux.ibm.com>