pkg/instance: fix 386 repros/jobs
All repros and jobs that use syz rerpos are broken for linux/amd64/386 since Mar 2022 with:

SYZFAIL: failed to recv rpc
aborting RPC server: mismatching manager/executor arches: amd64 vs 386

The problem is that ExecProgInstance.RunSyzProgFile passes sysTarget.Arch
(which is VMArch) to syz-execprog, so it expects executor to be amd64.

Restore the orignal logic.